requirements.txt
# Core ETL
apache-airflow==2.8.1
apache-airflow-providers-amazon==8.13.0
apache-airflow-providers-snowflake==5.2.1

# API & HTTP
requests==2.31.0
urllib3==2.1.0

# Data processing
pandas==2.1.4
pyarrow==14.0.2

# Cloud storage
boto3==1.34.14
s3fs==2023.12.2

# Snowflake
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
loguru==0.7.2
